worker: refactor client
